requirements.txt
pyswisseph==2.8.0.post1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
tzdata==2021.5
tzlocal==4.1
